on reflection, decided to separate out PM implementations to a separate file in case we add more in future - e.g. might want one that wraps the psutil library, one that gathers Java-specific stats, etc
